\name{initialize.smc}
\alias{initialize.smc}

\title{
    Initialize Sequential Monte Carlo
}

\description{

}

\usage{
initialize.smc(ws, ...)
}


\arguments{
    \item{ws}{
        smc.workspace object.
}
    \item{\dots}{}
}

\details{}

\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}

\references{
    Del Moral, P., Doucet, A., & Jasra, A. (2012) An adaptive sequential Monte Carlo method for approximate Bayesian computation. \emph{Statistics and Computing}, \bold{22}(5), 1009--1020.
}

\author{
    Art Poon and Mathias Renaud
}

\note{}

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}

\examples{

}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
