\name{simulate.trees}
\alias{simulate.trees}

\title{
    Simulate Trees Under a Model
}
\description{

}

\usage{
    simulate.trees(workspace, theta, seed = NA, ...)
}

\arguments{
    \item{workspace}{
        smc.workspace object.
}
    \item{theta}{
        parameter vector.
}
    \item{seed}{
        argument to \code{set.seed()}.
}
    \item{\dots}{}
}

\details{}

\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}

\references{
    Del Moral, P., Doucet, A., & Jasra, A. (2012) An adaptive sequential Monte Carlo method for approximate Bayesian computation. \emph{Statistics and Computing}, \bold{22}(5), 1009--1020.
}

\author{
    Art Poon and Mathias Renaud
}

\note{}

\seealso{

}

\examples{

}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }